separate main into controller and widget
