--- lib/main_screen.dart
@@ -1,65 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:pg_slema/features/diet/presentation/widget/diet_screen.dart';
-import 'package:pg_slema/features/exercises/presentation/widget/exercises_screen.dart';
-import 'package:pg_slema/features/motivation/presentation/widget/motivation_screen.dart';
-
-class MainScreen extends StatefulWidget {
-  const MainScreen({super.key});
-
-  @override
-  _MainScreenState createState() => _MainScreenState();
-}
-
-class _MainScreenState extends State<MainScreen> {
-  int _currentIndex = 0;
-  final PageController _pageController = PageController();
-
-  final List<Widget> _views = [
-    MotivationScreen(),
-    DietScreen(),
-    ExercisesScreen(),
-  ];
-
-  void _onTabTapped(int index) {
-    setState(() {
-      _currentIndex = index;
-      _pageController.animateToPage(
-        index,
-        duration: const Duration(milliseconds: 200),
-        curve: Curves.easeInOut,
-      );
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Main screen app bar'),
-      ),
-      body: PageView(
-        controller: _pageController,
-        children: _views,
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: _currentIndex,
-        iconSize: 40,
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.lightbulb),
-            label: 'Motivation',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.restaurant_menu),
-            label: 'Diet',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.fitness_center),
-            label: 'Exercises',
-          ),
-        ],
-        onTap: _onTabTapped,
-      ),
-    );
-  }
-}
