+++ test/widget_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:pg_slema/main.dart';
+import 'package:pg_slema/main/main.dart';
 
 void main() {
